vista_metodo_pago_categoria

+++ POLIRUBRO/capaNegocio/Verificar.cs
@@ -53,9 +53,9 @@ namespace POLIRUBRO
             else { MessageBox.Show("No se permite Stock negativo o Stock superior a 1000", "Stock",MessageBoxButtons.OK, MessageBoxIcon.Warning); return false; }
         }
 
-        public bool verificar_precio(int precio)
+        public bool verificar_precio(double precio)
         {
-            if (precio >0 && precio <= 1000000) { return true; }
+            if (precio >0 && precio <= 1000000.00) { return true; }
             else { MessageBox.Show("Precio debe ser mayor a 0 y menor que 1.00.0000","Precio",MessageBoxButtons.OK, MessageBoxIcon.Warning); return false; }
         }
